Populate custom fields automatically
It makes perfect sense that fields from `Additional attributes`
are populated automatically, instead of user having to hardcode
how the fields are populated in source code of this module.

This implementation has no support for translations or more
complex fields. But it should work fine with basic fields that
have simple value column.

The implementation is inspired by comment by @pablo-tapia in
AngelAlvarado/multiple_idp_simplesamlphp#4.