Auto-provisioning for existing users
Before, if an user with same email existed, new user with the
same email would be created anyways. But since we should be
able to trust added identity providers, we can say that they
are the same person and we can link the existing account
for the SSO user provided by IdP by adding a authmap entry.